src/requirements.txt
glfw
PyOpenGL
OpenARB
OpenGLU
OpenGLUT
Openshaders